+++ R/ps_arrange.R
@@ -36,6 +36,7 @@
 #' otu_table(pst)[1:8, 1:8]
 #' sample_data(pst) %>% head(8)
 ps_arrange <- function(ps, ..., .target = "sample_data") {
+  ps <- ps_get(ps)
   sample_order <- switch(.target,
     "sample_data" = {
       df <- data.frame(phyloseq::sample_data(ps))